backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
PyJWT==2.8.0
werkzeug==3.0.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
gunicorn==22.0.0
dnspython==2.6.1
python-dotenv==1.0.1
